Change dev layers
Delete /out
Change module-info
Add SceneChanger.java
Refactoring Service & Controller layers
