+++ src/main/resources/fxml/CopyPage.fxml
@@ -46,7 +46,7 @@
         </VBox.margin>
     </HBox>
     <HBox alignment="CENTER" spacing="10">
-        <Button onAction="#copyFiles" text="Start Copying"/>
+        <Button onAction="#copyButtonAction" text="Start Copying"/>
         <VBox.margin>
             <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
         </VBox.margin>
